moving landing componenets to landing folder
