+++ src/components/landing/OrbsInBoxes.tsx
@@ -1,7 +1,7 @@
 import { HighlightedText } from 'components/Text'
 import Colors, { colorToTailwindBg } from 'models/Colors'
 import OrbBox from 'icons/OrbBox'
-import ZkSphere from 'components/ZkSphere'
+import ZkSphere from 'components/landing/ZkSphere'
 import classnames, {
   TMargin,
   alignItems,
