+++ src/pages/Landing.tsx
@@ -2,15 +2,15 @@ import { HighlightedText } from 'components/Text'
 import BuildingIdentitiesCard from 'components/landing/BuildingIdentitiesCard'
 import Colors, { colorToTailwindBg } from 'models/Colors'
 import CreatingZKProofCard from 'components/landing/CreatingZKProofCard'
-import IdentityCard from 'components/IdentityCardOne'
+import IdentityCard from 'components/landing/IdentityCard'
 import InitialCard from 'components/landing/InitialCard'
 import LearnMoreCard from 'components/landing/LearnMoreCard'
-import OrbsInBoxes from 'components/OrbsInBoxes'
-import ScrollDownButton from 'components/ScrollDownButton'
+import OrbsInBoxes from 'components/landing/OrbsInBoxes'
+import ScrollDownButton from 'components/landing/ScrollDownButton'
 import SuperHr from 'icons/SuperHr'
 import SuperOrbWithConnectors from 'icons/SuperOrbWithConnectors'
 import TopConnectors from 'icons/TopConnectors'
-import ZkSphere from 'components/ZkSphere'
+import ZkSphere from 'components/landing/ZkSphere'
 import classnames, {
   alignItems,
   backgroundColor,
